Optimize UI, and add new feature!!!That's good.

+++ src/main/java/com/yichen/community/controller/PublishController.java
@@ -35,7 +35,7 @@ public class PublishController {
             @RequestParam(value = "title", required = false) String title,
             @RequestParam(value = "description", required = false) String description,
             @RequestParam(value = "tag", required = false) String tag,
-            @RequestParam(value = "id", required = false) Integer id,
+            @RequestParam(value = "id", required = false) Long id,
             HttpServletRequest httpServletRequest,
             Model model) {
 
@@ -76,7 +76,7 @@ public class PublishController {
     }
 
     @GetMapping("/publish/{id}")
-    public String rePublish(@PathVariable(value = "id") Integer id,
+    public String rePublish(@PathVariable(value = "id") Long id,
                             Model model) {
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
